requirements.txt
# Core ML
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=1.0.2
tensorflow>=2.8.0
opencv-python>=4.5.5
Pillow>=9.0.0

# Web Application
flask>=2.0.1
flask-cors>=3.0.10
gunicorn>=20.1.0
python-dotenv>=0.19.2

# Data Processing
matplotlib>=3.4.3
seaborn>=0.11.2

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1
locust>=2.8.4

# Utilities
tqdm>=4.62.3
python-dotenv>=0.19.2
python-multipart>=0.0.5

# Containerization
docker>=5.0.3

# Jupyter
jupyter>=1.0.0
ipykernel>=6.4.1
